refactor: 그룹 생성 및 업데이트 페이지 컴포넌트 분리 및 업데이트 상태시 초기 로딩 추가
